Update folder structure and refactor namespaces
