+++ TollCalculator/Interfaces/IVehicle.cs
@@ -1,0 +1,7 @@
+﻿namespace TollFeeCalculator
+{
+    public interface IVehicle
+    {
+        String GetVehicleType();
+    }
+}
