+++ TollCalculator/Models/Car.cs
@@ -1,0 +1,10 @@
+﻿namespace TollFeeCalculator
+{
+    public class Car : IVehicle
+    {
+        public String GetVehicleType()
+        {
+            return "Car";
+        }
+    }
+}
