+++ TollCalculator/Models/Motorbike.cs
@@ -1,0 +1,10 @@
+﻿namespace TollFeeCalculator
+{
+    public class Motorbike : IVehicle
+    {
+        public string GetVehicleType()
+        {
+            return "Motorbike";
+        }
+    }
+}
