--- TollCalculator/Motorbike.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TollFeeCalculator
-{
-    public class Motorbike : Vehicle
-    {
-        public string GetVehicleType()
-        {
-            return "Motorbike";
-        }
-    }
-}
